Stage_2/requirements.txt
keras==2.9.0
keras_nightly==2.10.0.dev2022071707
matplotlib==3.1.3
numpy==1.21.6
opencv_python==4.5.5.64
packaging==20.1
pandas==1.0.1
Pillow==9.2.0
scikit_image==0.16.2
scikit_learn==1.1.1
skimage==0.0
tensorflow==2.9.1
tf_nightly==2.10.0.dev20220717
Pillow==9.2.0
